First Draft of PT Fields (Json format)
First Draft of "PTFields Collect" and "PTFields Expand".  They take the main gen fields  (prompts, size, steps, etc.) and converts the to and from json format. These can be used with the prompts to file and from file nodes to read and write to files. 
Still need proper testing and validation adding to what happens if fields are missing from the file.